feat: implement workspace catalog endpoint, enhance RBAC permission management, and add frontend support for workspace initialization and role creation

+++ controlplane/internal/iam/domain/repo/user_repo.go
@@ -11,7 +11,7 @@ import (
 // [COMMENT]: UserRepository định nghĩa interface cho các thao tác quản lý dữ liệu danh bạ người dùng (User Directory)
 type UserRepository interface {
 	// [COMMENT]: ListUsers lấy danh sách users có level thấp hơn caller level (role_level số lớn hơn)
-	ListUsers(ctx context.Context, callerLevel int32, limit int, offset int) ([]*iamEntity.User, error)
+	ListUsers(ctx context.Context, callerLevel uint8, limit int, offset int) ([]*iamEntity.User, error)
 	// [COMMENT]: UpdateUserStatus cập nhật trạng thái hoạt động (status) của user dưới DB
 	UpdateUserStatus(ctx context.Context, userID uuid.UUID, status string) error
 	// [COMMENT]: GetUserProfile lấy thông tin profile hiển thị của user từ bảng user_profiles
